Extract Redis locks and pending claims into src/core/locks.ts
Move lock token helpers, withRedisLock/withVerificationActionLock, and
pending-claim state management into a dedicated module. Typecheck, 167
tests, and build all pass.
